automatically backed up by learn

+++ guessing_cli.rb
@@ -4,7 +4,7 @@ def run_guessing_game
     num = 1 + rand(6)
     puts "Guess a number between 1 and 6."
     input = gets.chomp
-    if input.to_f == num
+    if input.to_i == num
       puts "You guessed the correct number!"
     elsif input == "exit"
       puts "Goodbye!"
